+++ src/extension/Cycode.VisualStudio.Extension.Shared/Cli/DTO/ScanResult/Secret/SecretScanResult.cs
@@ -3,6 +3,5 @@
 namespace Cycode.VisualStudio.Extension.Shared.Cli.DTO.ScanResult.Secret;
 
 public class SecretScanResult: ScanResultBase {
-    public new List<SecretDetection> Detections { get; set; }
-    public new List<CliError> Errors { get; set; }
+    public List<SecretDetection> Detections { get; set; }
 }