+++ src/extension/Cycode.VisualStudio.Extension.Shared/Commands/CycodeToolWindowCommand.cs
@@ -1,8 +1,0 @@
-﻿namespace Cycode.VisualStudio.Extension.Shared;
-
-[Command(PackageIds.CycodeCommand)]
-internal sealed class CycodeToolWindowCommand : BaseCommand<CycodeToolWindowCommand> {
-    protected override Task ExecuteAsync(OleMenuCmdEventArgs e) {
-        return CycodeToolWindow.ShowAsync();
-    }
-}