+++ src/extension/Cycode.VisualStudio.Extension.Shared/Commands/RunAllScansCommand.cs
@@ -1,0 +1,19 @@
+﻿using Cycode.VisualStudio.Extension.Shared.DTO;
+using Cycode.VisualStudio.Extension.Shared.Services;
+
+namespace Cycode.VisualStudio.Extension.Shared;
+
+[Command(PackageIds.ToolbarRunAllScansCommand)]
+internal sealed class RunAllScansCommand : BaseCommand<RunAllScansCommand> {
+    protected override async Task ExecuteAsync(OleMenuCmdEventArgs e) {
+        IStateService stateService = ServiceLocator.GetService<IStateService>();
+        ExtensionState pluginState = stateService.Load();
+        if (!pluginState.CliAuthed) {
+            await VS.StatusBar.ShowMessageAsync("Please authenticate with Cycode first.");
+            return;
+        }
+
+        ICycodeService cycode = ServiceLocator.GetService<ICycodeService>();
+        await cycode.StartSecretScanForCurrentProjectAsync();
+    }
+}