+++ src/extension/Cycode.VisualStudio.Extension.Shared/Services/ErrorList/IErrorTaskCreatorService.cs
@@ -2,4 +2,5 @@
 
 public interface IErrorTaskCreatorService {
     Task RecreateAsync();
+    Task ClearErrorsAsync();
 }