+++ src/extension/Cycode.VisualStudio.Extension.Shared/VSCommandTable.cs
@@ -12,8 +12,14 @@ internal sealed class PackageGuids {
 ///     Helper class that encapsulates all CommandIDs uses across VS Package.
 /// </summary>
 internal sealed class PackageIds {
-    public const int CycodeCommand = 0x0100;
-    public const int OpenSettingsCommand = 0x0111;
-    public const int TWindowToolbar = 0x1000;
-    public const int TWindowToolbarGroup = 0x1050;
+    public const int ViewOpenToolWindowCommand = 0x1001;
+
+    public const int TWindowToolbar = 0x1050;
+    public const int ToolbarOpenSettingsCommand = 0x1052;
+    public const int ToolbarRunAllScansCommand = 0x1053;
+    public const int ToolbarOpenWebDocsCommand = 0x1054;
+    public const int ToolbarClearScanResultsCommand = 0x1055;
+
+    public const int TopMenuCycodeCommand = 0x1103;
+    public const int TopMenuOpenSettingsCommand = 0x1104;
 }