+++ src/extension/Cycode.VisualStudio.Extension.Shared/source.extension.cs
@@ -1,11 +1,7 @@
 namespace Cycode.VisualStudio.Extension.Shared;
 
 internal sealed class Vsix {
-    public const string Id = "Cycode.f2c5020e-67a2-46f8-a888-609412fd59db";
     public const string Name = "Cycode";
-    public const string Description = @"The Cycode Visual Studio Extension is a plugin for Visual Studio IDE";
-    public const string Language = "en-US";
-    public const string Version = "0.0.0";
-    public const string Author = "Cycode";
-    public const string Tags = "";
+    public const string Description = "Cycode for Visual Studio IDE";
+    public const string Version = "1.0.0";
 }