+++ CHANGELOG.md
@@ -1,3 +1,13 @@
 <!-- Keep a Changelog guide -> https://keepachangelog.com -->
 
-# Cycode plugin for Visual Studio
+# Cycode Visual Studio Extension
+
+## [Unreleased]
+
+## [1.0.0] - 2024-06-28
+
+The first public release of the extension.
+
+[1.0.0]: https://github.com/cycodehq/visual-studio-extension/releases/tag/v1.0.0
+
+[Unreleased]: https://github.com/cycodehq/visual-studio-extension/compare/v1.0.0...HEAD